requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
astroid==2.12.9
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.1
cleo==1.0.0a4
click==8.1.3
colorama==0.4.5
crashtest==0.3.1
Deprecated==1.2.13
dill==0.3.5.1
frozenlist==1.3.1
hiredis==2.0.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
multidict==6.0.2
packaging==21.3
platformdirs==2.5.2
pptree==3.1
pydantic==1.10.2
pylev==1.4.0
pylint==2.15.2
pylint-runner==0.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-ulid==1.1.0
pytz==2022.2.1
redis==4.3.4
redis-om==0.0.27
six==1.16.0
tomli==2.0.1
tomlkit==0.11.4
types-redis==4.3.20
types-six==1.16.19
typing_extensions==4.3.0
wrapt==1.14.1
yarl==1.8.1
